Completed simulation tests for the posts module
