+++ app/params/weights.go
@@ -1,6 +1,10 @@
 package params
 
-// Default simulation operation weights for messages and gov proposals
+// Default simulation operation weights for messages
 const (
-	DefaultWeightMsgCreatePost int = 100
+	DefaultWeightMsgCreatePost     int = 100
+	DefaultWeightMsgEditPost       int = 100
+	DefaultWeightMsgAddReaction    int = 100
+	DefaultWeightMsgRemoveReaction int = 100
+	DefaultWeightMsgAnswerPoll     int = 100
 )
