+++ x/posts/internal/types/msgs.go
@@ -1,7 +1,6 @@
 package types
 
 import (
-	"encoding/json"
 	"fmt"
 	"strings"
 	"time"
@@ -29,7 +28,8 @@ type MsgCreatePost struct {
 
 // NewMsgCreatePost is a constructor function for MsgSetName
 func NewMsgCreatePost(message string, parentID PostID, allowsComments bool, subspace string,
-	optionalData map[string]string, owner sdk.AccAddress, creationDate time.Time, medias PostMedias, pollData *PollData) MsgCreatePost {
+	optionalData map[string]string, owner sdk.AccAddress, creationDate time.Time,
+	medias PostMedias, pollData *PollData) MsgCreatePost {
 	return MsgCreatePost{
 		Message:        message,
 		ParentID:       parentID,
@@ -41,13 +41,6 @@ func NewMsgCreatePost(message string, parentID PostID, allowsComments bool, subs
 		Medias:         medias,
 		PollData:       pollData,
 	}
-}
-
-// MarshalJSON implements the custom marshaling as Amino does not support
-// the JSON signature omitempty
-func (msg MsgCreatePost) MarshalJSON() ([]byte, error) {
-	type msgCreatePost MsgCreatePost
-	return json.Marshal(msgCreatePost(msg))
 }
 
 // Route should return the name of the module
@@ -185,162 +178,4 @@ func (msg MsgEditPost) GetSignBytes() []byte {
 // GetSigners defines whose signature is required
 func (msg MsgEditPost) GetSigners() []sdk.AccAddress {
 	return []sdk.AccAddress{msg.Editor}
-}
-
-// ----------------------
-// --- MsgAnswerPoll
-// ----------------------
-
-// MsgAnswerPoll defines the AnswerPoll message
-type MsgAnswerPoll struct {
-	PostID      PostID         `json:"post_id"`
-	UserAnswers []uint         `json:"answers"`
-	Answerer    sdk.AccAddress `json:"answerer"`
-}
-
-// NewMsgAnswerPoll is the constructor function for MsgAnswerPoll
-func NewMsgAnswerPoll(id PostID, providedAnswers []uint, answerer sdk.AccAddress) MsgAnswerPoll {
-	return MsgAnswerPoll{
-		PostID:      id,
-		UserAnswers: providedAnswers,
-		Answerer:    answerer,
-	}
-}
-
-// Route should return the name of the module
-func (msg MsgAnswerPoll) Route() string { return RouterKey }
-
-// Type should return the action
-func (msg MsgAnswerPoll) Type() string { return ActionAnswerPoll }
-
-// ValidateBasic runs stateless checks on the message
-func (msg MsgAnswerPoll) ValidateBasic() error {
-	if !msg.PostID.Valid() {
-		return sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "Invalid post id")
-	}
-
-	if msg.Answerer.Empty() {
-		return sdkerrors.Wrap(sdkerrors.ErrInvalidAddress, fmt.Sprintf("Invalid answerer address: %s", msg.Answerer))
-	}
-
-	if len(msg.UserAnswers) == 0 {
-		return sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "Provided answers must contains at least one answer")
-	}
-
-	return nil
-}
-
-// GetSignBytes encodes the message for signing
-func (msg MsgAnswerPoll) GetSignBytes() []byte {
-	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(msg))
-}
-
-// GetSigners defines whose signature is required
-func (msg MsgAnswerPoll) GetSigners() []sdk.AccAddress { return []sdk.AccAddress{msg.Answerer} }
-
-// ----------------------
-// --- MsgAddPostReaction
-// ----------------------
-
-// MsgAddPostReaction defines the message to be used to add a reaction to a post
-type MsgAddPostReaction struct {
-	PostID PostID         `json:"post_id"` // Id of the post to react to
-	Value  string         `json:"value"`   // Reaction of the reaction
-	User   sdk.AccAddress `json:"user"`    // Address of the user reacting to the post
-}
-
-// NewMsgAddPostReaction is a constructor function for MsgAddPostReaction
-func NewMsgAddPostReaction(postID PostID, value string, user sdk.AccAddress) MsgAddPostReaction {
-	return MsgAddPostReaction{
-		PostID: postID,
-		User:   user,
-		Value:  value,
-	}
-}
-
-// Route should return the name of the module
-func (msg MsgAddPostReaction) Route() string { return RouterKey }
-
-// Type should return the action
-func (msg MsgAddPostReaction) Type() string { return ActionAddPostReaction }
-
-// ValidateBasic runs stateless checks on the message
-func (msg MsgAddPostReaction) ValidateBasic() error {
-	if !msg.PostID.Valid() {
-		return sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "Invalid post id")
-	}
-
-	if msg.User.Empty() {
-		return sdkerrors.Wrap(sdkerrors.ErrInvalidAddress, fmt.Sprintf("Invalid user address: %s", msg.User))
-	}
-
-	if len(strings.TrimSpace(msg.Value)) == 0 {
-		return sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, fmt.Sprintf("Reaction value cannot be empty nor blank"))
-	}
-
-	return nil
-}
-
-// GetSignBytes encodes the message for signing
-func (msg MsgAddPostReaction) GetSignBytes() []byte {
-	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(msg))
-}
-
-// GetSigners defines whose signature is required
-func (msg MsgAddPostReaction) GetSigners() []sdk.AccAddress {
-	return []sdk.AccAddress{msg.User}
-}
-
-// ----------------------
-// --- MsgRemovePostReaction
-// ----------------------
-
-// MsgRemovePostReaction defines the message to be used when wanting to remove
-// an existing reaction from a specific user having a specific value
-type MsgRemovePostReaction struct {
-	PostID   PostID         `json:"post_id"`  // Id of the post to unlike
-	User     sdk.AccAddress `json:"user"`     // Address of the user that has previously liked the post
-	Reaction string         `json:"reaction"` // Reaction of the reaction to be removed
-}
-
-// MsgUnlikePostPost is the constructor of MsgRemovePostReaction
-func NewMsgRemovePostReaction(postID PostID, user sdk.AccAddress, reaction string) MsgRemovePostReaction {
-	return MsgRemovePostReaction{
-		PostID:   postID,
-		User:     user,
-		Reaction: reaction,
-	}
-}
-
-// Route should return the name of the module
-func (msg MsgRemovePostReaction) Route() string { return RouterKey }
-
-// Type should return the action
-func (msg MsgRemovePostReaction) Type() string { return ActionRemovePostReaction }
-
-// ValidateBasic runs stateless checks on the message
-func (msg MsgRemovePostReaction) ValidateBasic() error {
-	if !msg.PostID.Valid() {
-		return sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "Invalid post id")
-	}
-
-	if msg.User.Empty() {
-		return sdkerrors.Wrap(sdkerrors.ErrInvalidAddress, fmt.Sprintf("Invalid user address: %s", msg.User))
-	}
-
-	if len(strings.TrimSpace(msg.Reaction)) == 0 {
-		return sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "Reaction value cannot be empty nor blank")
-	}
-
-	return nil
-}
-
-// GetSignBytes encodes the message for signing
-func (msg MsgRemovePostReaction) GetSignBytes() []byte {
-	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(msg))
-}
-
-// GetSigners defines whose signature is required
-func (msg MsgRemovePostReaction) GetSigners() []sdk.AccAddress {
-	return []sdk.AccAddress{msg.User}
 }
