+++ x/posts/internal/types/msgs_test.go
@@ -483,207 +483,3 @@ func TestMsgEditPost_GetSigners(t *testing.T) {
 	assert.Equal(t, 1, len(actual))
 	assert.Equal(t, msgEditPost.Editor, actual[0])
 }
-
-// ----------------------
-// --- MsgAddPostReaction
-// ----------------------
-
-var msgLike = types.NewMsgAddPostReaction(types.PostID(94), "like", testOwner)
-
-func TestMsgAddPostReaction_Route(t *testing.T) {
-	actual := msgLike.Route()
-	assert.Equal(t, "posts", actual)
-}
-
-func TestMsgAddPostReaction_Type(t *testing.T) {
-	actual := msgLike.Type()
-	assert.Equal(t, "add_post_reaction", actual)
-}
-
-func TestMsgAddPostReaction_ValidateBasic(t *testing.T) {
-	tests := []struct {
-		name  string
-		msg   types.MsgAddPostReaction
-		error error
-	}{
-		{
-			name:  "Invalid post id returns error",
-			msg:   types.NewMsgAddPostReaction(types.PostID(0), "like", testOwner),
-			error: sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "Invalid post id"),
-		},
-		{
-			name:  "Invalid user returns error",
-			msg:   types.NewMsgAddPostReaction(types.PostID(5), "like", nil),
-			error: sdkerrors.Wrap(sdkerrors.ErrInvalidAddress, "Invalid user address: "),
-		},
-		{
-			name:  "Invalid value returns error",
-			msg:   types.NewMsgAddPostReaction(types.PostID(5), "", testOwner),
-			error: sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "Reaction value cannot be empty nor blank"),
-		},
-		{
-			name:  "Valid message returns no error",
-			msg:   types.NewMsgAddPostReaction(types.PostID(10), "like", testOwner),
-			error: nil,
-		},
-	}
-
-	for _, test := range tests {
-		test := test
-		returnedError := test.msg.ValidateBasic()
-		if test.error == nil {
-			assert.Nil(t, returnedError)
-		} else {
-			assert.NotNil(t, returnedError)
-			assert.Equal(t, test.error.Error(), returnedError.Error())
-		}
-	}
-}
-
-func TestMsgAddPostReaction_GetSignBytes(t *testing.T) {
-	actual := msgLike.GetSignBytes()
-	expected := `{"type":"desmos/MsgAddPostReaction","value":{"post_id":"94","user":"cosmos1cjf97gpzwmaf30pzvaargfgr884mpp5ak8f7ns","value":"like"}}`
-	assert.Equal(t, expected, string(actual))
-}
-
-func TestMsgAddPostReaction_GetSigners(t *testing.T) {
-	actual := msgLike.GetSigners()
-	assert.Equal(t, 1, len(actual))
-	assert.Equal(t, msgLike.User, actual[0])
-}
-
-// ----------------------
-// --- MsgRemovePostReaction
-// ----------------------
-
-var msgUnlikePost = types.NewMsgRemovePostReaction(types.PostID(94), testOwner, "like")
-
-func TestMsgUnlikePost_Route(t *testing.T) {
-	actual := msgUnlikePost.Route()
-	assert.Equal(t, "posts", actual)
-}
-
-func TestMsgUnlikePost_Type(t *testing.T) {
-	actual := msgUnlikePost.Type()
-	assert.Equal(t, "remove_post_reaction", actual)
-}
-
-func TestMsgUnlikePost_ValidateBasic(t *testing.T) {
-	tests := []struct {
-		name  string
-		msg   types.MsgRemovePostReaction
-		error error
-	}{
-		{
-			name:  "Invalid post id returns error",
-			msg:   types.NewMsgRemovePostReaction(types.PostID(0), testOwner, "like"),
-			error: sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "Invalid post id"),
-		},
-		{
-			name:  "Invalid user address: ",
-			msg:   types.NewMsgRemovePostReaction(types.PostID(10), nil, "like"),
-			error: sdkerrors.Wrap(sdkerrors.ErrInvalidAddress, "Invalid user address: "),
-		},
-		{
-			name:  "Invalid value returns no error",
-			msg:   types.NewMsgRemovePostReaction(types.PostID(10), testOwner, ""),
-			error: sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "Reaction value cannot be empty nor blank"),
-		},
-		{
-			name:  "Valid message returns no error",
-			msg:   types.NewMsgRemovePostReaction(types.PostID(10), testOwner, "like"),
-			error: nil,
-		},
-	}
-
-	for _, test := range tests {
-		returnedError := test.msg.ValidateBasic()
-		if test.error == nil {
-			assert.Nil(t, returnedError)
-		} else {
-			assert.NotNil(t, returnedError)
-			assert.Equal(t, test.error.Error(), returnedError.Error())
-		}
-	}
-}
-
-func TestMsgUnlikePost_GetSignBytes(t *testing.T) {
-	actual := msgUnlikePost.GetSignBytes()
-	expected := `{"type":"desmos/MsgRemovePostReaction","value":{"post_id":"94","reaction":"like","user":"cosmos1cjf97gpzwmaf30pzvaargfgr884mpp5ak8f7ns"}}`
-	assert.Equal(t, expected, string(actual))
-}
-
-func TestMsgUnlikePost_GetSigners(t *testing.T) {
-	actual := msgUnlikePost.GetSigners()
-	assert.Equal(t, 1, len(actual))
-	assert.Equal(t, msgUnlikePost.User, actual[0])
-}
-
-// ----------------------
-// --- MsgAnswerPoll
-// ----------------------
-
-var msgAnswerPollPost = types.NewMsgAnswerPoll(types.PostID(1), []uint{1, 2}, testOwner)
-
-func TestMsgAnswerPollPost_Route(t *testing.T) {
-	actual := msgAnswerPollPost.Route()
-	assert.Equal(t, "posts", actual)
-}
-
-func TestMsgAnswerPollPost_Type(t *testing.T) {
-	actual := msgAnswerPollPost.Type()
-	assert.Equal(t, "answer_poll", actual)
-}
-
-func TestMsgAnswerPollPost_ValidateBasic(t *testing.T) {
-	tests := []struct {
-		name  string
-		msg   types.MsgAnswerPoll
-		error error
-	}{
-		{
-			name:  "Invalid post id",
-			msg:   types.NewMsgAnswerPoll(types.PostID(0), []uint{1, 2}, msgAnswerPollPost.Answerer),
-			error: sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "Invalid post id"),
-		},
-		{
-			name:  "Invalid answerer address",
-			msg:   types.NewMsgAnswerPoll(types.PostID(1), []uint{1, 2}, nil),
-			error: sdkerrors.Wrap(sdkerrors.ErrInvalidAddress, "Invalid answerer address: "),
-		},
-		{
-			name:  "Returns error when no answer is provided",
-			msg:   types.NewMsgAnswerPoll(types.PostID(1), []uint{}, msgAnswerPollPost.Answerer),
-			error: sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "Provided answers must contains at least one answer"),
-		},
-		{
-			name: "Valid message returns no error",
-			msg:  types.NewMsgAnswerPoll(types.PostID(1), []uint{1, 2}, msgAnswerPollPost.Answerer),
-		},
-	}
-
-	for _, test := range tests {
-		test := test
-		t.Run(test.name, func(t *testing.T) {
-			returnedError := test.msg.ValidateBasic()
-			if test.error == nil {
-				assert.Nil(t, returnedError)
-			} else {
-				assert.NotNil(t, returnedError)
-				assert.Equal(t, test.error.Error(), returnedError.Error())
-			}
-		})
-	}
-}
-
-func TestMsgAnswerPollPost_GetSignBytes(t *testing.T) {
-	actual := msgAnswerPollPost.GetSignBytes()
-	expected := `{"type":"desmos/MsgAnswerPoll","value":{"answerer":"cosmos1cjf97gpzwmaf30pzvaargfgr884mpp5ak8f7ns","answers":["1","2"],"post_id":"1"}}`
-	assert.Equal(t, expected, string(actual))
-}
-
-func TestMsgAnswerPollPost_GetSigners(t *testing.T) {
-	actual := msgAnswerPollPost.GetSigners()
-	assert.Equal(t, 1, len(actual))
-	assert.Equal(t, msgAnswerPollPost.Answerer, actual[0])
-}
